Bring available vars for auto-broadcast into parity with chat commands
With this commit, all message variables which are available in chat commands are now also available in auto broadcast messages.
Additionally, vars which were previously only available in auto broadcast, are now also available in chat commands.
